backend/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
urllib3==2.0.7
Werkzeug==2.3.0
Flask==2.3.1
Flask-API==3.1
Flask-APScheduler==1.13.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
boto3==1.29.2
botocore==1.32.2
Babel==2.13.1
bcrypt==4.0.1
gunicorn==21.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
passlib==1.7.4
python-dotenv==1.0.1